Reduce unit testing
